requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
weaviate-client>=3.26.7,<4.0.0
openai==1.98.0
python-dotenv==1.1.1
tiktoken==0.9.0
numpy==2.3.2
pandas==2.3.1
psutil==7.0.0
matplotlib==3.10.5
seaborn==0.13.2
tqdm==4.67.1